type change

+++ src/main/config/config.c
@@ -390,7 +390,7 @@ uint32_t getPidUpdateRate(void) {
     }
 }
 
-uint16_t getGyroUpdateRate(void) {
+uint32_t getGyroUpdateRate(void) {
     return gyro.targetLooptime;
 }
